Consolidate task UI
